refactored the code the reduce repetitions
