+++ client/src/pages/Accounts/Signup/SignUp.tsx
@@ -1,18 +1,11 @@
-import CssBaseline from '@material-ui/core/CssBaseline';
-import Paper from '@material-ui/core/Paper';
-import Box from '@material-ui/core/Box';
-import Grid from '@material-ui/core/Grid';
-import { Link } from 'react-router-dom';
 import { FormikHelpers } from 'formik';
-import Typography from '@material-ui/core/Typography';
-import useStyles from '../useStyles';
+import AccountWrapper from '../../../components/AccountWrapper/AccountWrapper';
 import register from '../../../helpers/APICalls/register';
 import SignUpForm from '../AuthForm/SignupForm/SignUpForm';
 import { useAuth } from '../../../context/useAuthContext';
 import { useSnackBar } from '../../../context/useSnackbarContext';
 
 export default function Register(): JSX.Element {
-  const classes = useStyles();
   const { updateLoginContext } = useAuth();
   const { updateSnackBarMessage } = useSnackBar();
 
@@ -38,26 +31,8 @@ export default function Register(): JSX.Element {
   };
 
   return (
-    <Grid container justify="center" component="main" className={classes.root}>
-      <CssBaseline />
-      <Grid item elevation={6} className={classes.addMargin} component={Paper} square>
-        <Box className={classes.authWrapper}>
-          <Box width="100%" maxWidth={450} p={3} alignSelf="center">
-            <Grid container>
-              <Grid item xs>
-                <Typography align="center" className={classes.welcome} component="h1" variant="h5">
-                  Sign up
-                </Typography>
-              </Grid>
-            </Grid>
-            <SignUpForm handleSubmit={handleSubmit} />
-            <Typography align="center" className={classes.redirect}>
-              Already a member?
-              <Link to="./login">Login</Link>
-            </Typography>
-          </Box>
-        </Box>
-      </Grid>
-    </Grid>
+    <AccountWrapper>
+      <SignUpForm handleSubmit={handleSubmit} />
+    </AccountWrapper>
   );
 }
